refactor: move remote_function logic to make a space for other types (#322)
